+++ changelog.md
@@ -1,3 +1,6 @@
 - Switch Rune of Resurrection to LootTables/JSON
+- Added config for the Rune of Locked Time stopping falling blocks
+- Added config blacklist for rune of level earth
+- Added config whitelist for farmland replaceable blocks 
 
 Uses Runes of Wizardry 0.6.5